requirement.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
mysql==0.0.3
mysqlclient==2.2.5
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
